+++ R/plots-qc-filtering.R
@@ -282,7 +282,7 @@ plot_qc_summary_byclass <- function(
 #' @details
 #' The QC filtering process follows a hierarchical structure, where features are first evaluated against lower-level filters such as signal-to-blank ratios and limit of detection (LOD).
 #' Only features that pass these basic criteria are then subjected to higher-level filters like the coefficient of variation (CV) or linear regression results.
-#' A feature will only fail a higher-level filter (such as `CV` or `R-squared`) if it has passed all previous lower-level filters.
+#' A feature will only fail a higher-level filter (such as `CV` or `R²`) if it has passed all previous lower-level filters.
 #' This ensures that features are evaluated progressively, starting from fundamental quality checks up to more stringent filtering criteria.
 #'
 #' Note: The function currently shows a warning `Using `size` aesthetic for lines was deprecated in ggplot2 3.4.0.` which can be ignored.